+++ src/Ext/Scenario/Body.cpp
@@ -81,8 +81,8 @@ void ScenarioExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 
 	// INI_EX exINI(pINI);
 
-
-
+	ParTitle = ScenarioClass::Instance->OverParTitle;
+	ParMessage = ScenarioClass::Instance->OverParMessage;
 }
 
 template <typename T>
@@ -92,6 +92,10 @@ void ScenarioExt::ExtData::Serialize(T& Stm)
 		.Process(this->Waypoints)
 		.Process(this->Variables[0])
 		.Process(this->Variables[1])
+		.Process(this->ParTitle)
+		.Process(this->ParMessage)
+		.Process(this->ScoreCampaignTheme)
+		.Process(this->NextMission)
 		;
 }