+++ src/Ext/Scenario/Body.h
@@ -24,10 +24,18 @@ public:
 	public:
 		std::map<int, CellStruct> Waypoints;
 		std::map<int, ExtendedVariable> Variables[2]; // 0 for local, 1 for global
+		CSFText ParTitle;
+		CSFText ParMessage;
+		Nullable<PhobosFixedString<0x20>> ScoreCampaignTheme;
+		Nullable<PhobosFixedString<0x20>> NextMission;
 
 		ExtData(ScenarioClass* OwnerObject) : Extension<ScenarioClass>(OwnerObject)
 			, Waypoints { }
 			, Variables { }
+			, ParTitle { nullptr }
+			, ParMessage { nullptr }
+			, ScoreCampaignTheme { }
+			, NextMission { }
 		{ }
 
 		void SetVariableToByID(bool bIsGlobal, int nIndex, char bState);