+++ src/Misc/PhobosGlobal.cpp
@@ -1,0 +1,54 @@
+#include "PhobosGlobal.h"
+#include <Ext/WarheadType/Body.h>
+
+PhobosGlobal PhobosGlobal::GlobalObject;
+
+PhobosGlobal* PhobosGlobal::Global()
+{
+	return &GlobalObject;
+}
+
+void PhobosGlobal::Clear()
+{
+	GlobalObject.Reset();
+}
+
+void PhobosGlobal::Reset()
+{
+	RandomTriggerPool.clear();
+}
+
+//Save/Load
+#pragma region save/load
+
+template <typename T>
+bool PhobosGlobal::Serialize(T& stm)
+{
+	return stm
+		.Process(this->RandomTriggerPool)
+		.Success();
+}
+
+bool PhobosGlobal::Save(PhobosStreamWriter& stm)
+{
+	return Serialize(stm);
+}
+
+bool PhobosGlobal::Load(PhobosStreamReader& stm)
+{
+	return Serialize(stm);
+}
+
+bool PhobosGlobal::SaveGlobals(PhobosStreamWriter& stm)
+{
+	GlobalObject.Save(stm);
+	return stm.Success();
+}
+
+bool PhobosGlobal::LoadGlobals(PhobosStreamReader& stm)
+{
+	GlobalObject.Load(stm);
+	return stm.Success();
+}
+
+#pragma endregion save/load