+++ docs/Whats-New.md
@@ -52,6 +52,10 @@ You can use the migration utility (can be found on [Phobos supplementaries repo]
   58=Upper bound,0
   59=Operate var is global,10
   60=Operate var index,0
+  61=RandomPoolID,0
+  62=RemoveFromPool,10
+  63=IsTitle,10
+  64=Map,8
 
   [EventsRA2]
   500=Local variable is greater than,48,6,0,0,[LONG DESC],0,1,500,1
@@ -101,6 +105,12 @@ You can use the migration utility (can be found on [Phobos supplementaries repo]
   504=Binary operation,0,56,55,60,54,59,0,0,0,[LONG DESC],0,1,504,1
   505=Fire Super Weapon at specified location (Phobos),0,0,20,2,21,22,0,0,0,Launch a Super Weapon from [SuperWeaponTypes] list at a specified location. House=-1 means random target that isn't neutral. House=-2 means the first neutral house. House=-3 means random human target. Coordinate X=-1 means random. Coordinate Y=-1 means random,0,1,505
   506=Fire Super Weapon at specified waypoint (Phobos),0,0,20,2,30,0,0,0,0,Launch a Super Weapon from [SuperWeaponTypes] list at a specified waypoint. House=-1 means random target that isn't neutral. House=-2 means the first neutral house. House=-3 means random human target. Coordinate X=-1 means random. Coordinate Y=-1 means random,0,1,506
+  12000=Put Trigger into specificed Pool,-2,14,61,0,0,0,0,0,0,Add trigger to specified random pool enable by action 12002,0,1,12000,1
+  12001=Remove a Trigger from specified Pool,-2,14,61,0,0,0,0,0,0,Remove trigger from specificed random pool,0,1,12001,1
+  12002=Enable a Trigger of specified Pool,0,0,61,62,0,0,0,0,0,Randomized enable a trigger of specified Random Pool,0,1,12002,1
+  19000=Set Text of Score Interface ,-4,13,63,0,0,0,0,0,0,Set text of score interface Title or Message as Text of CSFLabel,0,1,19000,1
+  19001=Set Theme of Score Interface,-4,17,0,0,0,0,0,0,0,Set theme of score interface,0,1,19001,1
+  19002=Set Next Mission,-4,64,0,0,0,0,0,0,0,Set next mission. Parm is ID in mapselmd.ini,0,1,19002,1
 
   ; FOLLOWING ENTRIES REQUIRE FA2SP.DLL (by secsome)
   [ScriptTypeLists]
@@ -333,6 +343,9 @@ New:
 - Implemented support for PCX images for campaign loading screen (by FlyStar)
 - Implemented support for PCX images for observer loading screen (by Uranusian)
 - Animated (non-tiberium spawning) TerrainTypes (by Starkku)
+- Trigger Action 9932-9934 Random Trigger Pool (by ststl)
+- Trigger Action 9935,9936 set text, Theme of score interface (by FlyStar)
+- Trigger Action 9937 set next mission (by FlyStar)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)